Give ComposerBar the status as data, not as an Element
ComposerBar took badges and status as Option<Element>. An untyped slot is
an invitation for each surface to put something different in it, and both
took it: the agent chat passed a real status with its subagent, task and
queue counts, and the start page passed a hardcoded span reading "Ready".
Start is always ready. The label carried nothing.

The status is four fields now, and the component decides what to draw
from them. Nothing to say draws nothing, which is what start passes and
why its label is gone without a flag saying so. ComposerStatus and the
StatusDot it needs move to vmux_ui so both surfaces render the same
component rather than two that had drifted; the agent header picks up the
same dot. Its nine hardcoded strings become Fluent keys on the way, and
composer-ready leaves the locales with the span that used it.
