+++ crates/page/vmux_chat/src/page/agent.rs
@@ -1,6 +1,7 @@
 use super::state::Chat;
 use dioxus::prelude::*;
 use vmux_ui::back::BackButton;
+use vmux_ui::components::composer_bar::StatusDot;
 use vmux_ui::favicon::favicon_src_for_url;
 
 #[component]
@@ -58,21 +59,6 @@ fn AgentAvatar(chat: Chat, size_class: String) -> Element {
                 "{initial}"
             }
         }
-    }
-}
-
-#[component]
-pub fn StatusDot(status: String, size_class: String) -> Element {
-    let tone = match status.as_str() {
-        "streaming" => "bg-amber-400 shadow-[0_0_8px_rgba(251,191,36,0.65)]",
-        "installing" => "bg-sky-400 shadow-[0_0_8px_rgba(56,189,248,0.65)]",
-        "awaiting" => "bg-violet-400 shadow-[0_0_8px_rgba(167,139,250,0.65)]",
-        "interrupted" => "bg-amber-400 shadow-[0_0_8px_rgba(251,191,36,0.65)]",
-        "errored" => "bg-red-500 shadow-[0_0_8px_rgba(239,68,68,0.65)]",
-        _ => "bg-success shadow-[0_0_8px_rgba(16,185,129,0.65)]",
-    };
-    rsx! {
-        span { class: "{size_class} rounded-full {tone}" }
     }
 }
