+++ crates/vmux_ui/locales/ar.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = إزالة المرفق
 composer-choose-agent = اختر وكيلاً
 composer-create-worktree = + شجرة عمل
 composer-create-worktree-title = إنشاء أو اختيار شجرة عمل لهذا المشروع
+composer-status-running = قيد التشغيل
+composer-status-approval = بانتظار الموافقة
+composer-status-starting = يبدأ
+composer-status-error = خطأ
+composer-status-ready = جاهز
+composer-status-subagents = وكلاء فرعيون نشطون
+composer-status-tasks-title = مهام مفتوحة
+composer-status-tasks = { $count } مهام
+composer-status-queued-title = مطالبات في الانتظار
+composer-status-queued = { $count } في الانتظار
 composer-uncommitted-changes = تغييرات غير مُودعة
 composer-commits-ahead = إيداعات متقدمة على المصدر
 composer-no-git = بدون Git
-composer-ready = جاهز
 composer-git = Git
 composer-git-repository = مستودع Git
 composer-branch-name = الفرع { $branch }
