+++ crates/vmux_ui/locales/de.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = Anhang entfernen
 composer-choose-agent = Agent auswählen
 composer-create-worktree = + Worktree
 composer-create-worktree-title = Worktree für dieses Projekt erstellen oder auswählen
+composer-status-running = Läuft
+composer-status-approval = Freigabe
+composer-status-starting = Startet
+composer-status-error = Fehler
+composer-status-ready = Bereit
+composer-status-subagents = Aktive Subagenten
+composer-status-tasks-title = Offene Aufgaben
+composer-status-tasks = { $count } Aufgaben
+composer-status-queued-title = Wartende Prompts
+composer-status-queued = { $count } in Warteschlange
 composer-uncommitted-changes = Nicht committete Änderungen
 composer-commits-ahead = Commits vor dem Upstream
 composer-no-git = Kein Git
-composer-ready = Bereit
 composer-git = Git
 composer-git-repository = Git-Repository
 composer-branch-name = Branch { $branch }
