+++ crates/vmux_ui/locales/en-US.ftl
@@ -273,10 +273,19 @@ composer-remove-attachment = Remove attachment
 composer-choose-agent = Choose agent
 composer-create-worktree = + Worktree
 composer-create-worktree-title = Create or select a worktree for this project
+composer-status-running = Running
+composer-status-approval = Approval
+composer-status-starting = Starting
+composer-status-error = Error
+composer-status-ready = Ready
+composer-status-subagents = Active subagents
+composer-status-tasks-title = Open plan tasks
+composer-status-tasks = { $count } tasks
+composer-status-queued-title = Queued prompts
+composer-status-queued = { $count } queued
 composer-uncommitted-changes = Uncommitted changes
 composer-commits-ahead = Commits ahead of upstream
 composer-no-git = No Git
-composer-ready = Ready
 composer-git = Git
 composer-git-repository = Git repository
 composer-branch-name = Branch { $branch }
