+++ crates/vmux_ui/locales/es.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = Quitar adjunto
 composer-choose-agent = Elegir agente
 composer-create-worktree = + Worktree
 composer-create-worktree-title = Crear o seleccionar un worktree para este proyecto
+composer-status-running = En curso
+composer-status-approval = Aprobación
+composer-status-starting = Iniciando
+composer-status-error = Error
+composer-status-ready = Listo
+composer-status-subagents = Subagentes activos
+composer-status-tasks-title = Tareas abiertas
+composer-status-tasks = { $count } tareas
+composer-status-queued-title = Prompts en cola
+composer-status-queued = { $count } en cola
 composer-uncommitted-changes = Cambios sin confirmar
 composer-commits-ahead = Commits por delante del upstream
 composer-no-git = Sin Git
-composer-ready = Listo
 composer-git = Git
 composer-git-repository = Repositorio Git
 composer-branch-name = Rama { $branch }
