+++ crates/vmux_ui/locales/fr.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = Supprimer la pièce jointe
 composer-choose-agent = Choisir un agent
 composer-create-worktree = + Worktree
 composer-create-worktree-title = Créer ou sélectionner un worktree pour ce projet
+composer-status-running = En cours
+composer-status-approval = Approbation
+composer-status-starting = Démarrage
+composer-status-error = Erreur
+composer-status-ready = Prêt
+composer-status-subagents = Sous-agents actifs
+composer-status-tasks-title = Tâches ouvertes
+composer-status-tasks = { $count } tâches
+composer-status-queued-title = Prompts en file
+composer-status-queued = { $count } en file
 composer-uncommitted-changes = Modifications non validées
 composer-commits-ahead = Commits en avance sur l’amont
 composer-no-git = Pas de Git
-composer-ready = Prêt
 composer-git = Git
 composer-git-repository = Dépôt Git
 composer-branch-name = Branche { $branch }
