+++ crates/vmux_ui/locales/it.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = Rimuovi allegato
 composer-choose-agent = Scegli agente
 composer-create-worktree = + Worktree
 composer-create-worktree-title = Crea o seleziona un worktree per questo progetto
+composer-status-running = In corso
+composer-status-approval = Approvazione
+composer-status-starting = Avvio
+composer-status-error = Errore
+composer-status-ready = Pronto
+composer-status-subagents = Subagenti attivi
+composer-status-tasks-title = Attività aperte
+composer-status-tasks = { $count } attività
+composer-status-queued-title = Prompt in coda
+composer-status-queued = { $count } in coda
 composer-uncommitted-changes = Modifiche non confermate
 composer-commits-ahead = Commit in anticipo sull’upstream
 composer-no-git = Nessun Git
-composer-ready = Pronto
 composer-git = Git
 composer-git-repository = Repository Git
 composer-branch-name = Ramo { $branch }
