+++ crates/vmux_ui/locales/ja.ftl
@@ -236,10 +236,19 @@ composer-remove-attachment = 添付ファイルを削除
 composer-choose-agent = エージェントを選択
 composer-create-worktree = + ワークツリー
 composer-create-worktree-title = このプロジェクトのワークツリーを作成または選択
+composer-status-running = 実行中
+composer-status-approval = 承認待ち
+composer-status-starting = 起動中
+composer-status-error = エラー
+composer-status-ready = 待機中
+composer-status-subagents = 稼働中のサブエージェント
+composer-status-tasks-title = 未完了のタスク
+composer-status-tasks = タスク { $count } 件
+composer-status-queued-title = キュー中のプロンプト
+composer-status-queued = キュー { $count } 件
 composer-uncommitted-changes = コミットされていない変更
 composer-commits-ahead = アップストリームより先行するコミット
 composer-no-git = Git なし
-composer-ready = 準備完了
 composer-git = Git
 composer-git-repository = Git リポジトリ
 composer-branch-name = ブランチ { $branch }
