+++ crates/vmux_ui/locales/ko.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = 첨부 파일 삭제
 composer-choose-agent = 에이전트 선택
 composer-create-worktree = + 워크트리
 composer-create-worktree-title = 이 프로젝트의 워크트리를 만들거나 선택
+composer-status-running = 실행 중
+composer-status-approval = 승인 대기
+composer-status-starting = 시작 중
+composer-status-error = 오류
+composer-status-ready = 준비됨
+composer-status-subagents = 활성 서브에이전트
+composer-status-tasks-title = 열린 작업
+composer-status-tasks = 작업 { $count }개
+composer-status-queued-title = 대기 중인 프롬프트
+composer-status-queued = 대기 { $count }개
 composer-uncommitted-changes = 커밋되지 않은 변경 사항
 composer-commits-ahead = 업스트림보다 앞선 커밋
 composer-no-git = Git 없음
-composer-ready = 준비됨
 composer-git = Git
 composer-git-repository = Git 저장소
 composer-branch-name = 브랜치 { $branch }
