+++ crates/vmux_ui/locales/nl.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = Bijlage verwijderen
 composer-choose-agent = Agent kiezen
 composer-create-worktree = + Worktree
 composer-create-worktree-title = Een worktree voor dit project maken of kiezen
+composer-status-running = Actief
+composer-status-approval = Goedkeuring
+composer-status-starting = Starten
+composer-status-error = Fout
+composer-status-ready = Gereed
+composer-status-subagents = Actieve subagents
+composer-status-tasks-title = Openstaande taken
+composer-status-tasks = { $count } taken
+composer-status-queued-title = Prompts in wachtrij
+composer-status-queued = { $count } in wachtrij
 composer-uncommitted-changes = Niet-vastgelegde wijzigingen
 composer-commits-ahead = Commits voor op upstream
 composer-no-git = Geen Git
-composer-ready = Gereed
 composer-git = Git
 composer-git-repository = Git-repository
 composer-branch-name = Branch { $branch }
