+++ crates/vmux_ui/locales/pl.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = Usuń załącznik
 composer-choose-agent = Wybierz agenta
 composer-create-worktree = + Worktree
 composer-create-worktree-title = Utwórz lub wybierz worktree dla tego projektu
+composer-status-running = Działa
+composer-status-approval = Zatwierdzenie
+composer-status-starting = Uruchamianie
+composer-status-error = Błąd
+composer-status-ready = Gotowe
+composer-status-subagents = Aktywne subagenty
+composer-status-tasks-title = Otwarte zadania
+composer-status-tasks = Zadania: { $count }
+composer-status-queued-title = Prompty w kolejce
+composer-status-queued = W kolejce: { $count }
 composer-uncommitted-changes = Niezatwierdzone zmiany
 composer-commits-ahead = Commity przed gałęzią nadrzędną
 composer-no-git = Brak Git
-composer-ready = Gotowe
 composer-git = Git
 composer-git-repository = Repozytorium Git
 composer-branch-name = Gałąź { $branch }
