+++ crates/vmux_ui/locales/pt-BR.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = Remover anexo
 composer-choose-agent = Escolher agente
 composer-create-worktree = + Worktree
 composer-create-worktree-title = Criar ou selecionar um worktree para este projeto
+composer-status-running = Em execução
+composer-status-approval = Aprovação
+composer-status-starting = Iniciando
+composer-status-error = Erro
+composer-status-ready = Pronto
+composer-status-subagents = Subagentes ativos
+composer-status-tasks-title = Tarefas abertas
+composer-status-tasks = { $count } tarefas
+composer-status-queued-title = Prompts na fila
+composer-status-queued = { $count } na fila
 composer-uncommitted-changes = Alterações não confirmadas
 composer-commits-ahead = Commits à frente do upstream
 composer-no-git = Sem Git
-composer-ready = Pronto
 composer-git = Git
 composer-git-repository = Repositório Git
 composer-branch-name = Ramo { $branch }
