+++ crates/vmux_ui/locales/ru.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = Удалить вложение
 composer-choose-agent = Выбрать агента
 composer-create-worktree = + Worktree
 composer-create-worktree-title = Создать или выбрать рабочее дерево для этого проекта
+composer-status-running = Выполняется
+composer-status-approval = Подтверждение
+composer-status-starting = Запуск
+composer-status-error = Ошибка
+composer-status-ready = Готово
+composer-status-subagents = Активные субагенты
+composer-status-tasks-title = Открытые задачи
+composer-status-tasks = Задач: { $count }
+composer-status-queued-title = Промпты в очереди
+composer-status-queued = В очереди: { $count }
 composer-uncommitted-changes = Незафиксированные изменения
 composer-commits-ahead = Коммиты впереди upstream
 composer-no-git = Без Git
-composer-ready = Готово
 composer-git = Git
 composer-git-repository = Репозиторий Git
 composer-branch-name = Ветка { $branch }
