+++ crates/vmux_ui/locales/uk.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = Прибрати вкладення
 composer-choose-agent = Вибрати агента
 composer-create-worktree = + Worktree
 composer-create-worktree-title = Створити або вибрати робоче дерево для цього проєкту
+composer-status-running = Виконується
+composer-status-approval = Підтвердження
+composer-status-starting = Запуск
+composer-status-error = Помилка
+composer-status-ready = Готово
+composer-status-subagents = Активні субагенти
+composer-status-tasks-title = Відкриті завдання
+composer-status-tasks = Завдань: { $count }
+composer-status-queued-title = Промпти в черзі
+composer-status-queued = У черзі: { $count }
 composer-uncommitted-changes = Незафіксовані зміни
 composer-commits-ahead = Коміти попереду upstream
 composer-no-git = Без Git
-composer-ready = Готово
 composer-git = Git
 composer-git-repository = Репозиторій Git
 composer-branch-name = Гілка { $branch }
