+++ crates/vmux_ui/locales/zh-Hans.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = 移除附件
 composer-choose-agent = 选择代理
 composer-create-worktree = + 工作树
 composer-create-worktree-title = 为此项目创建或选择工作树
+composer-status-running = 运行中
+composer-status-approval = 待批准
+composer-status-starting = 启动中
+composer-status-error = 错误
+composer-status-ready = 就绪
+composer-status-subagents = 活动子代理
+composer-status-tasks-title = 未完成任务
+composer-status-tasks = { $count } 项任务
+composer-status-queued-title = 排队的提示
+composer-status-queued = 排队 { $count } 项
 composer-uncommitted-changes = 未提交的更改
 composer-commits-ahead = 领先上游的提交
 composer-no-git = 无 Git
-composer-ready = 就绪
 composer-git = Git
 composer-git-repository = Git 仓库
 composer-branch-name = 分支 { $branch }
