+++ crates/vmux_ui/locales/zh-Hant.ftl
@@ -248,10 +248,19 @@ composer-remove-attachment = 移除附件
 composer-choose-agent = 選擇代理
 composer-create-worktree = + 工作樹
 composer-create-worktree-title = 為此專案建立或選擇工作樹
+composer-status-running = 執行中
+composer-status-approval = 待核准
+composer-status-starting = 啟動中
+composer-status-error = 錯誤
+composer-status-ready = 就緒
+composer-status-subagents = 作用中的子代理
+composer-status-tasks-title = 未完成工作
+composer-status-tasks = { $count } 項工作
+composer-status-queued-title = 排隊的提示
+composer-status-queued = 排隊 { $count } 項
 composer-uncommitted-changes = 未提交的變更
 composer-commits-ahead = 領先上游的提交
 composer-no-git = 無 Git
-composer-ready = 就緒
 composer-git = Git
 composer-git-repository = Git 儲存庫
 composer-branch-name = 分支 { $branch }
